feat: разбивает main.js на модули
