+++ js/constants.js
@@ -1,0 +1,57 @@
+export const NAMES = [
+  'Иван',
+  'Мария',
+  'Алексей',
+  'Елена',
+  'Дмитрий',
+  'Анна',
+  'Сергей',
+  'Ольга',
+  'Максим',
+  'Екатерина'
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const DESCRIPTIONS = [
+  'Закат на пляже',
+  'Горный пейзаж',
+  'Городская архитектура',
+  'Портрет друга',
+  'Натюрморт с фруктами',
+  'Дождливый день',
+  'Семейное фото',
+  'Концерт в парке',
+  'Прогулка по лесу',
+  'Вечер в кафе',
+  'Море и волны',
+  'Зимний пейзаж',
+  'Цветы в саду',
+  'Уличное искусство',
+  'Котёнок играет',
+  'Рассвет в горах',
+  'Ночной город',
+  'Пикник на природе',
+  'Осенние листья',
+  'Спортивное мероприятие',
+  'Винтажная машина',
+  'Романтический ужин',
+  'Детская площадка',
+  'Музей изнутри',
+  'Танцы на площади'
+];
+
+export const CONFIG = {
+  PHOTOS_COUNT: 25,
+  LIKES_RANGE: {min: 15, max: 200},
+  COMMENTS_RANGE: {min: 0, max: 30},
+  MESSAGE_COUNT_RANGE: {min: 1, max: 2},
+  AVATAR_RANGE: {min: 1, max: 6}
+};
